feat: 添加角色和聊天服务API，更新相关逻辑以支持对话请求参数解析，优化对话处理功能

+++ backend/services/chat/api/internal/logic/chat/getconversationhistorylogic.go
@@ -32,7 +32,7 @@ func (l *GetConversationHistoryLogic) GetConversationHistory(req *types.GetConve
 	chatRepo := repo.NewChatServiceRepo(l.ctx, l.svcCtx)
 
 	// 获取对话历史
-	conversations, total, err := chatRepo.GetConversationsByUserID(req.UserID, req.Page, req.PageSize, req.CharacterID)
+	conversations, total, err := chatRepo.GetConversationsByUserID(req.UserID, req.Page, req.PageSize, int64(req.CharacterID))
 	if err != nil {
 		l.Logger.Error("GetConversationHistory failed: ", err)
 		return nil, err
